make validate function common
